pyproject.toml
[tool.poetry]
name = "job_agent"
version = "0.1.0"
description = "AI-powered job application assistant"
authors = ["Yanying Gu <yanying.gu@gmail.com>"]
readme = "README.md"
packages = [{include = "job_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
python-frontmatter = "^1.0.0"
# Add your dependencies here as you use them
# Example:
# openai = "^1.0.0"
# langchain = "^0.1.0"
# anthropic = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^23.0"
ruff = "^0.1.0"
mypy = "^1.0"
notebook = "^7.0"  # Use notebook instead of jupyter to avoid jupyterlab dependency
ipykernel = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=job_agent --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "tests/",
    "notebooks/",
]
